Add Larastan package to composer dependencies
A new package, Larastan, has been added to the development dependencies in the composer.json file. This PHPStan wrapper for Laravel will help in identifying and fixing issues in the code base without running it. The corresponding changes in composer.lock file demonstrate the addition of Larastan and its dependencies.