Update Post model and add publish/unpublish notifications
The 'meta_description' cast type in the Post model has been changed from 'array' to 'json'. Furthermore, successful publish and unpublish actions for posts in the EditPost page now trigger corresponding notifications. These notifications provide instant feedback to the user, indicating the completion of the publish/unpublish process along with the new status of the post.